+++ NEWS.md
@@ -4,7 +4,7 @@
 * new `.combine_value_labels` argument for `recode.haven_labelled()` (#61)
 * new `drop_unused_value_labels()` method
 * an additional `.labels` argument for `set_value_labels()`
-* `user_na_to_na` argument has been to `to_character.haven_labelled()`
+* `user_na_to_na` argument has been added to `to_character.haven_labelled()`
 * `%>%` is now imported from `dplyr`
 * a cheatsheet has been added (#47)