+++ NEWS.md
@@ -1,3 +1,5 @@
+# labelled (development version)
+
 # labelled 2.5.0
 
 * new `.keep_value_labels` argument for `recode.haven_labelled()`