+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `set_variable_labels`, `set_value_labels`, `add_value_labels`,
   and `remove_value_labels` now accept "tidy dots"
+* new function `names_prefixed_by_values()` to get the names of a vector
+  prefixed by their corresponding value
 
 # labelled 2.5.0