+++ NEWS.md
@@ -10,7 +10,7 @@
       simpler character vectors
     - `generate_dictionary()` is an equivalent of `look_for()`
 * `set_variable_labels`, `set_value_labels`, `add_value_labels`,
-  and `remove_value_labels` now accept "tidy dots"
+  and `remove_value_labels` now accept "tidy dots" (#67 @psanker)
 * new function `names_prefixed_by_values()` to get the names of a vector
   prefixed by their corresponding value